Web3 login design changes (#601)
* setup

Moved Web2 and Web3 login into components

* toggle web2 login

* using Form

* toggle refactor

* dropdown

* h3 in components + mobile

* Update front-end/src/ui-components/AddressDropdown.tsx


* adding Loader

* Loader dimmer height
