+++ front-end/src/ui-components/GlobalStyle.tsx
@@ -69,12 +69,16 @@ export const GlobalStyle = createGlobalStyle`
         margin-top: 0;
     }
 
-    h1, h2 {
+    h1, h2, h3 {
         font-weight: 400;
         font-size: xl;
     }
 
-    h3, h4, h5, h6 {
+    h3 {
+        font-weight: 400;
+    }
+
+    h4, h5, h6 {
         font-weight: 500; 
     }
